refactor : static 상수들 constants 패키지 아래로 분리
