+++ src/main/java/javalotto/constants/LottoConstants.java
@@ -1,0 +1,10 @@
+package javalotto.constants;
+
+public final class LottoConstants {
+
+    public static final int LOTTO_NUMBER_MIN_VALUE = 1;
+    public static final int LOTTO_NUMBER_MAX_VALUE = 45;
+    public static final int LOTTO_NUMBERS_COUNT = 6;
+
+    private LottoConstants(){};
+}
