+++ src/main/java/javalotto/constants/PurchaseAmountConstants.java
@@ -1,0 +1,10 @@
+package javalotto.constants;
+
+public final class PurchaseAmountConstants {
+
+    public static final int PURCHASE_AMOUNT_MIN_VALUE = 1000;
+    public static final int PURCHASE_AMOUNT_MAX_VALUE = Integer.MAX_VALUE / 1000 * 1000;
+    public static final int PURCHASE_AMOUNT_UNIT_PRICE = 1000;
+
+    private PurchaseAmountConstants() {}
+}
