+++ src/main/java/javalotto/domain/LottoGenerator.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static javalotto.domain.Lotto.*;
+import static javalotto.constants.LottoConstants.*;
 
 public class LottoGenerator {
