+++ src/main/java/javalotto/domain/PurchaseAmount.java
@@ -5,10 +5,9 @@ import javalotto.exception.purchaseamount.AmountOutOfRangeException;
 
 import java.util.Objects;
 
+import static javalotto.constants.PurchaseAmountConstants.*;
+
 public class PurchaseAmount {
-    public static final int PURCHASE_AMOUNT_MIN_VALUE = 1000;
-    public static final int PURCHASE_AMOUNT_MAX_VALUE = Integer.MAX_VALUE / 1000 * 1000;
-    public static final int PURCHASE_AMOUNT_UNIT_PRICE = 1000;
 
     private final int purchaseAmount;
